Introduce general Performance class to handle events
