+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -11,6 +11,7 @@ import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.feature.Feature;
 import seedu.address.model.performance.Event;
+import seedu.address.model.performance.Performance;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
@@ -151,10 +152,10 @@ public class ParserUtil {
     public static Event parseEvent(String event) throws ParseException {
         requireNonNull(event);
         String trimmedEvent = event.trim();
-        if (!Event.doesExist(trimmedEvent)) {
+        if (!Performance.doesEventExist(trimmedEvent)) {
             throw new ParseException(String.format(Event.MESSAGE_CONSTRAINTS, trimmedEvent));
         }
-        return Event.getEvent(trimmedEvent);
+        return Performance.retrieveEvent(trimmedEvent);
     }
 
     /**
